requirements.txt
beautifulsoup4
mechanize
Pillow
keras
tensorflow
